pyproject.toml
[project]
name = "axiom"
version = "0.1.0"
description = "FastAPI benchmarking application for LLM agent evaluation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.0.0",
    "openai-agents>=0.0.7",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "loguru>=0.7.3",
    "plotperfect @ git+https://github.com/NASA-IMPACT/plotperfect.git",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
